+++ src/Modules/Profile/components/ProfileMenu.vue
@@ -88,7 +88,7 @@
         <span v-html="$i18n('profile.nav.violations', { count: violationCount })" />
       </b-list-group-item>
       <b-list-group-item
-        v-if="fsId !== fsIdSession && buttonNameReportRequest"
+        v-if="showReportButton"
         type="button"
         class="list-group-item list-group-item-action"
         href="#"
@@ -97,7 +97,7 @@
         <i class="far fa-life-ring fa-fw" /> {{ buttonNameReportRequest }}
       </b-list-group-item>
       <b-list-group-item
-        v-if="fsId !== fsIdSession"
+        v-if="showModerationButton"
         type="button"
         class="list-group-item list-group-item-action"
         href="#"
@@ -107,6 +107,7 @@
       </b-list-group-item>
     </b-list-group>
     <b-modal
+      v-if="showModerationButton"
       ref="modal_mediation"
       :title="$i18n('profile.mediation.title', { name: foodSaverName })"
       :cancel-title="$i18n('button.cancel')"
@@ -119,6 +120,7 @@
       />
     </b-modal>
     <b-modal
+      v-if="showReportButton"
       ref="modal_report_request"
       :title="$i18n('profile.report.title', { name: foodSaverName })"
       :cancel-title="$i18n('button.cancel')"
@@ -187,6 +189,14 @@ export default {
     return {
       isBuddy: false,
     }
+  },
+  computed: {
+    showReportButton () {
+      return this.buttonNameReportRequest !== null && this.buttonNameReportRequest.length > 0
+    },
+    showModerationButton () {
+      return this.fsId !== this.fsIdSession
+    },
   },
   methods: {
     openChat (fsId) {